requirements.txt
yt-dlp
python-dotenv
python-telegram-bot==13.12